Updated SQL schema-generated domain objects matching the latest schema rev from Laura and Daniella.  I also created a repository (DAO) for every domain object, even though some might never get used.  The ToolLibrary<->AnalysisTool relation is implemented, properly annotated, and has an MVC controller mapping and seems to work as before using the test JavaScript.  Hopefully all the rest of it yokes up as easily.
Notes:

Must set fetch=FetchType.EAGER (at the field level of domain objects) in the @OneToMany annotation of the domain object, in order for automatic(when desired) forward traversal of @OneToMany relations in the result object. This is at the data access level.  This is apparently a Hibernate integration issue since Hibernate lazy-loads these relations by default.  There is a Jackson component that is supposed to handle this but we can't use it because of an incompatibility with the current Spring version (3.1.x as opposed to 3.2.x). Anyway it seems to work fine when fetch type is set to eager, so I'm moving on.

Also, *must* make sure to do proper @Json annotations on relationships (@JsonBackReference, @JsonForwardReference, @JsonIgnore) in SQL-generated domain objects or there will definitely be errors when serialzing JSON in the autowired MVC context. Jackson serialization can go into infinite loops when traversing self-referential objects, unless they are properly annotated. This affects only the web server, but can be a PITA to detect when it happens, as client just gets 500 error after server spazzes out.